+++ CHANGELOG.md
@@ -651,3 +651,4 @@ This is a fast forward to v1.6.3 of daviscook477's fork with a few additional ch
   * Meditate
   * Flicker (deprecated)
 * Fix many actions to work with max hand size changes (kiooeht)
+* Fix bias in boss selection RNG (dbjorge)